requirements.txt
# FastAPI 및 웹 서버
fastapi
uvicorn

# 데이터베이스
sqlalchemy
alembic

# 인증 및 보안
passlib[bcrypt]
python-jose[cryptography]
python-multipart

# 환경 변수 및 설정
python-dotenv

# HTTP 요청
requests
httpx

# AI 및 NLP
google-generativeai
sentence-transformers
konlpy

# 웹 스크래핑
beautifulsoup4
readability

# 데이터 처리
pandas
numpy

# 검증 및 스키마
pydantic[email]
email-validator

# 로깅 및 모니터링
python-json-logger

# 기타 유틸리티
pyparsing
typing-extensions

